Rename beaconNodeObservers.ts to observer.ts
