+++ packages/beacon-node/src/chain/archiveStore/observers/archiveBlocksObserver.ts
@@ -1,12 +1,12 @@
 import {BeaconConfig} from "@lodestar/config";
 import {CheckpointWithHex, IForkChoice} from "@lodestar/fork-choice";
 import {Logger} from "@lodestar/logger";
-import {QueueObserver} from "../../../beaconNodeObservers.js";
 import {IBeaconDb} from "../../../db/interface.js";
 import {IClock} from "../../../util/clock.js";
 import {LightClientServer} from "../../lightClient/index.js";
 import {PROCESS_FINALIZED_CHECKPOINT_QUEUE_LEN} from "../constants.js";
 import {archiveBlocks} from "../utils/archiveBlocks.js";
+import { QueueObserver } from "../../observer.js";
 
 export class ArchiveBlocksObserver extends QueueObserver {
   constructor(
