+++ packages/beacon-node/src/chain/archiveStore/observers/frequentStateArchiveObserver.ts
@@ -1,5 +1,5 @@
 import {CheckpointWithHex} from "@lodestar/fork-choice";
-import {QueueObserver} from "../../../beaconNodeObservers.js";
+import { QueueObserver } from "../../observer.js";
 import {PROCESS_FINALIZED_CHECKPOINT_QUEUE_LEN} from "../constants.js";
 import {FrequentStateArchiveModules, maybeArchiveState} from "../utils/frequentStateArchive.js";
