+++ packages/beacon-node/src/chain/archiveStore/observers/pruneUnfinalizedStateObserver.ts
@@ -2,7 +2,7 @@ import {CheckpointWithHex, IForkChoice} from "@lodestar/fork-choice";
 import {Logger} from "@lodestar/logger";
 import {CachedBeaconStateAllForks} from "@lodestar/state-transition";
 import {Checkpoint} from "@lodestar/types/lib/phase0/types.js";
-import {ChainObserver} from "../../../beaconNodeObservers.js";
+import { ChainObserver } from "../../observer.js";
 import {IStateRegenerator} from "../../regen/interface.js";
 
 export class PruneUnfinalizedStateObserver extends ChainObserver {
